+++ src/netbots_ipc.py
@@ -1,5 +1,4 @@
 import socket
-import umsgpack
 import random
 import time
 import re
@@ -7,6 +6,15 @@ import math
 import argparse
 
 from netbots_log import log
+
+import umsgpack
+# The code below breaks the netbots viewer. Only enable for better performance when no viewer will be used.
+#try:
+#    import msgpack as umsgpack
+#    log("Using binary python msgpack."
+#except:
+#    import umsgpack
+#    log("Using pure python msgpack. Install binary msgpack for better performance.", "WARNING")
 
 """
 **About Messages**